requirements.txt
# requirements.txt
python-dotenv
openpyxl
deepgram-sdk
elevenlabs
groq
livekit
sounddevice
soundfile
httpx
numpy
